+++ src/renderer/Skybox.h
@@ -1,6 +1,10 @@
 #pragma once
 #include <string>
-#include <OpenGL/gl3.h>
+#ifdef __APPLE__
+  #include <OpenGL/gl3.h>
+#else
+  #include <glad/glad.h>
+#endif
 
 class Camera;